refactor project file set up
