+++ src/components/Coin/Coin.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Coin = () => <h4>Coin</h4>;
+
+export default Coin;
