+++ src/components/CoinList/CoinList.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Coin from '../Coin/Coin';
+
+const CoinList = () => (
+  <div>
+    <h4>CoinList</h4>
+    <Coin />
+  </div>
+);
+
+export default CoinList;
